(SERVER-2172) Move x509Loader class to own file
